[CHANGE] Update Mission Cog
